Built airflow DAG in python to create a extracting, tranformation and loading process
Airflow, Docker and PostgreSQL was configured to be used. Airflow DAG recalls_etl_v1 was created successfully. But it failed when it was run on airflow DAGs due to ERROR - Failed to execute job 10 for task get_raw_data (Unterminated string starting at: line 1 column 3244243 (char 3244242); 2199). Will continue to fix this issue.